Transfer user function to user controller, define env mail
